apps/storage/pyproject.toml
[tool.poetry]
name = "storage_service"
version = "0.1.0"
description = ""
authors = ["olatejulian <julian.lf.olate@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.78.0"
SQLAlchemy = "^1.4.36"
python-interface = "^1.6.1"
python-dotenv = "^0.20.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
